Fix the ColdBootVisit race condition Part 2
We are reassigning the webview delegate to the session via the default
implementation of the sessionDidLoadWebView: method
(https://github.com/turbolinks/turbolinks-ios/blob/swift-3.0/Turbolinks/Session.swift#L15).

That method is:

 * called by the Session's implementation of visitDidInitializeWebView: method
   (https://github.com/turbolinks/turbolinks-ios/blob/swift-3.0/Turbolinks/Session.swift#L154)
 * which is called by the ColdBootVisit implementation of completeVisit()
   (https://github.com/turbolinks/turbolinks-ios/blob/swift-3.0/Turbolinks/Session.swift#L154)
 *  which is called by the complete() method on Visit
    (https://github.com/turbolinks/turbolinks-ios/blob/swift-3.0/Turbolinks/Visit.swift#L70)
 * which is called by ColdBootVisit's implementation of webView:didLoadPageWithRestorationIdentifier:
   (https://github.com/turbolinks/turbolinks-ios/blob/swift-3.0/Turbolinks/Visit.swift#L236).

The race condition remains, and if the WKScriptMessageHandler fires before
webView:didFinish:, the framework will drop the callback
to sessionDidFinishRequest:

This commit fixes this by implementing webView:didFinish: inside the
Session, and calling the SessionDelegate callback here as well. This feels a little
hacky, and does not resolve the issue if the developer overrides the default
implementation of sessionDidLoadWebView: and steals the delegate themselves. I
also worry this could lead to the callback being triggered twice, but so far
it *seems* this is not possible. A more general fix for this issue may require
some larger refactoring.